refactor: классы связанные с чатом перемещены в одну папку
