+++ src/main/java/com/javaacademy/onegramchat/chat/MenuButton.java
@@ -1,4 +1,4 @@
-package com.javaacademy.onegramchat;
+package com.javaacademy.onegramchat.chat;
 
 public enum MenuButton {
     LOGIN(1),
